Move the chat completions mockers to a new file
